requirements.txt
# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# Base ------------------------------------------------------------------------
gitpython
ipython  # interactive notebook
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.7.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.8.1
tqdm>=4.64.0

# Logging ---------------------------------------------------------------------
tensorboard>=2.4.1


# Plotting --------------------------------------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export ----------------------------------------------------------------------
# coremltools>=6.0  # CoreML export
onnx>=1.12.0  # ONNX export
onnx-simplifier>=0.4.1  # ONNX simplifier
WebUI @ git+https://github.com/Widdershin/flask-desktop.git@9f0e9e033909516e1c0669fdf38e9c9c904fe7ac
yolov5==7.0.8
gdown